Add support for form based logins to ruTorrent
While HTTP Basic Auth is the most common, some installations use a
HTML based login form that sets cookies to login to ruTorrent.

To support these use cases, we look for the standardized browser
autocomplete hints in the HTML to identify which form elements we
should fill (the same way a browser password autofill would). We
then submit the form automatically.

Because we are relying on the autocomplete hints, we can support
thousands of login form permutations without writing code that
expects any particular type of form.